Install prebuilt tools for unbundled builds with old build projects
The previous patch stopped installing the prebuilt tools because a patch
to the build project references them here instead, but some unbundled
manifests use an old version of the build/ project that expects the
tools to be installed but the master version of this project.  Detect if
the build system expects the tools to be installed and install them if
necessary.

Change-Id: Id20fb2a1c44f4c5e67d842b1e6d2215107bb2d24