API 26: Rename Collections.singleton*() type arg E -> T
This rename is part of the update of Collections.java to OpenJDK8u60.

This is compatible change because generic type argument names affect
neither compile time nor runtime compatibility.

The build system (incorrectly) flags this as a change to a released
API. Removing or changing released public APIs is infeasible on AOSP.
Therefore this change will remain restricted to internal branches
until the API is released.

Bug: 31902309
Test: make

Change-Id: I71ce677245a571d7cee50e6d4699343e5221bf5c