DO NOT MERGE: Fix searching for overridden methods.
If the superclass is an abstract class, we want to search it for
possibly overridden methods; we don't want to search ourselves.

This bug resulted in (incorrectly) thinking that overridden "default"
methods from interfaces were standalone methods on the class.  (The
expected behavior has always been to omit any overridden methods from
the public API surface area to avoid redundant noise.)

Test: make update-api
Bug: 64219148
Change-Id: I3d1afb8aee1793851b097109c85f2a89bb6f0a90