Put system stub in 'system_N' directory when finalizing.
Temporary to work around a bug in the build system where an android_system.jar
will cause it to expect a system_N dir with android.jar in it.

These rule and files are going to change a fair bit in master, so this is the
easiest fix for now.

Bug: 77588754
Test: prebuilts/sdk/update_prebuilts.py -f 28 4721199 && test m -j sdk_v28 sdk_vsystem_28
Change-Id: I7c4f3e2b4edd6727ef3d717b17a6f8565a0cd63d